[RELEASE] Version 1.2.0 is out
This release ships a lot of improvements
for containers in TYPO3:

Much more flexible container configuration with Objects
used with "Registry->configureContainer()" instead of
"Registry->addContainer()". See an example
on how to migrate here:
https://github.com/b13/container-example/pull/9/files

Containers can now be individually grouped (TYPO3 v10 only)
during the registration so they can be grouped in the selection
dropdowns and are visually separated.
Thanks to Georg Ringer for the suggestions and improvements.

Full support for EXT:content_defender. Thanks to Nicole
Cordes for her great support.

A lot of bugfixes with multi-language / translations
and workspaces retrieval.

All changes are still fully backwards-compatible
but it is recommended to use the new registration
as this will be the new way for a possible v2 release.