[TASK] Add guzzlehttp/psr7:^1.8.2 as v11 test dependency
codeception 4.1.21 to 4.1.22 allows guzzlehttp/psr7:^2.0
which was previously blocked, so some 1.x version came in.

See
https://github.com/Codeception/Codeception/commit/405204b192cb54462acd5ce78e614c1a0b36fac9

core 11.2 has an issue with guzzlehttp/psr7:^2.0 that
makes functional tests hard fail. This has been fixed
with core 11.3.

core 11.3 however can not be installed yet, since
content_defender only allows 11.2, even in its master
version.

We thus have to continue to test with core 11.2 for now,
so we explicitly require youngest 1.x version of
guzzlehttp/psr7 to not trigger the core 11.2 issue.