[DOCS] Transform into full documentation
Prepare the documentation for publishing at docs.typo3.org and thus
becoming part of the large and growing TYPO3 compendium.

The single documentation has been split into the documentation for
the VCS host (README.md) and for docs.typo3.org (Documentation/).
All content except the abstract has been moved to reST files under
Documentation/, and conforms to the latest standards of basic TYPO3
documentation.

The README.md and the full documentation should both guide the reader
to all aspects of the extension, like the TER page (download &
installation), full documentation and the VCS repository.

Use the TER page as the central page of your extension, which then
takes the user further, e.g. to Packagist.