[FEATURE] delete records with non existing parent
add a new command to delete records which points
to an non existing container record.
This records are displayed in BE as unused elements.
There is no need to delete this elements, because an
editor can delete them in BE. But this command maybe helpful
when migrating to EXT:container from any other grid extension
and having a lot of unused records which are not shown before
migration.