[BUGFIX] container with content_defender maxitems
ColumnConfigurationManipulationHook of EXT:content_defender is used for
* new content elemment wizard
* TcaCTypeItems: new record
* TcaCTypeItems: edit record
in this cases we should not consider maxitems configuration of container colPos.
this leeds to errors, because relation to container gets lost in EXT:content_defender.
because EXT:container supressed already rendering of the 'New Content' Button in Backend
if the maxitems of colPos is reached, intervention of EXT:content_defender is not necessary.

Fixes: #334
Relates to: https://github.com/IchHabRecht/content_defender/issues/105