[BUGFIX] Allow arbitrary container colPos values
Instead of implicitly assuming that a non-integer colPos is in the
legacy "<container-uid>-<colPos>" format, we now explicitly check for the
existence of the legacy "-" delimiter.

This allows arbitrary colPos values for containers in TYPO3v12. Here the
colPos in markup is a string like "<container-uid><v12-delimiter><colPos>"
which can easily exceed the integer maximum value in PHP in case the
"<colPos>" from a container is large itself. E.g. if a UNIX timestamp is
used as colPos.