[TASK] Adopt codeception helper login page check from testing-framework
For the codeception acceptance tests a custom codeception helper
is used to verify the login state, using `/typo3/index.php` to
check for backend login mask page.

Using the full endpoint is not reasonable and breaks with TYPO3
v13.0 and newer which removed that endpoint completly and this
change modifes the check to use `/typo3` path instead, adopting
from provided `typo3/testing-framework` helper which is fully
backwards compatible.