[BUGFIX] custom backend templates
the container grid is not longer rendered after the
content of the StandardContentRenderer but is assign
as a variable "tx_container_grid" to the record.

this variable is used in a backend template:
<f:format.raw>{tx_container_grid}</f:format.raw>

container shipped the default backend template (again, we has
dropped this in EXT:container 3.0)

in v12 and v13 the ContainerPreviewRenderer is no longer used
we use PageContentPreviewRendering Event Listener instead of
the ContainerPreviewRenderer
the Listener assigns the container grid to tx_container_grid

in v11 the PreviewRenderer is still used but assigns the
grid as a variable instead of rendering the grid after the content

background: in v13 the Event Listener "FluidBasedContentPreviewRenderer"
was introduced for rendering backend templates and "disables" StandardContentRenderer
if a backend template is defined.

in addition with this change you can also render custom stuff
in you backend template after the grid, which was not possible before

Fixes: #563