[TASK] park current GridColumnItem in runtime cache
to not break the BeforeContainerPreviewIsRenderedEvent Method Signatures
we put the GridColumnItem into the runtime cache in the ContainerPreviewRenderer
renderPageModulePreviewHeader Method (which is always called) and fetch it in
the GridRenderer to pass it tho the Event.
This is a hack to not break the event, but fix the custom BackendTemplate usage
(s. dfc61c4a711c3de1b288e5a9514a35f0f30a7234)

Relates to: #563